Fix for review (#3910)
